Incorporate review comments
