+++ optaplanner-core/src/test/java/org/optaplanner/core/config/solver/SolverConfigMultiThreadedTest.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.optaplanner.core.config.solver;
 
 import java.util.List;
@@ -8,7 +24,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.optaplanner.core.api.solver.Solver;
 import org.optaplanner.core.api.solver.SolverFactory;
-import org.optaplanner.core.config.solver.testutil.TestThreadFactory;
+import org.optaplanner.core.config.solver.testutil.MockThreadFactory;
 import org.optaplanner.core.impl.testdata.domain.TestdataEntity;
 import org.optaplanner.core.impl.testdata.domain.TestdataSolution;
 import org.optaplanner.core.impl.testdata.domain.TestdataValue;
@@ -118,7 +134,7 @@ public class SolverConfigMultiThreadedTest {
     public void customThreadFactoryClassIsUsed() {
         SolverFactory<TestdataSolution> solverFactory = PlannerTestUtils.buildSolverFactory(
                 TestdataSolution.class, TestdataEntity.class);
-        solverFactory.getSolverConfig().setThreadFactoryClass(TestThreadFactory.class);
+        solverFactory.getSolverConfig().setThreadFactoryClass(MockThreadFactory.class);
         solverFactory.getSolverConfig().setMoveThreadCount("2");
         Solver<TestdataSolution> solver = solverFactory.buildSolver();
 
@@ -126,6 +142,6 @@ public class SolverConfigMultiThreadedTest {
         solution = solver.solve(solution);
 
         assertSolution(solver, solution);
-        assertTrue(TestThreadFactory.hasBeenCalled());
+        assertTrue(MockThreadFactory.hasBeenCalled());
     }
 }
